+++ h04-async-file-io-promise/async-file-io-promise.js
@@ -1,4 +1,4 @@
-const express = require('express');
+const express = require("express");
 const app = express();
 const fs = require("fs");
 
@@ -6,7 +6,9 @@ const dataDir = "./async-file-io-data";
 
 let nextRequestId = 1;
 
-function add(a, b) { return a + b; }
+function add(a, b) {
+    return a + b;
+}
 
 app.use(express.json());
 
@@ -16,12 +18,16 @@ app.use((req, res, next) => {
     next();
 });
 
-app.get('/hello', (req, res) => {
-    res.send('Hello World! now is ' + Date.now());
+app.get("/hello", (req, res) => {
+    res.send("Hello World! now is " + Date.now());
 });
 
 app.post("/write", (req, res) => {
-    console.log(`start processing received request nr. ${req.id} at ${getFormattedTime(new Date(Date.now()))}`);
+    console.log(
+        `start processing received request nr. ${req.id} at ${getFormattedTime(
+            new Date(Date.now()),
+        )}`,
+    );
     const numFiles = req.body.numFiles;
     const value = req.body.value;
     for (let fileNo = 1; fileNo <= numFiles; fileNo++) {
@@ -29,11 +35,19 @@ app.post("/write", (req, res) => {
         fs.writeFileSync(fileName, value);
     }
     res.status(200).send("done");
-    console.log(`answered request nr. ${req.id} at ${getFormattedTime(new Date(Date.now()))}\n`);
+    console.log(
+        `answered request nr. ${req.id} at ${getFormattedTime(
+            new Date(Date.now()),
+        )}\n`,
+    );
 });
 
 app.post("/writeSeq", (req, res) => {
-    console.log(`[WRITE-SEQ] start processing received request nr. ${req.id} at ${getFormattedTime(new Date(Date.now()))}`);
+    console.log(
+        `[WRITE-SEQ] start processing received request nr. ${
+            req.id
+        } at ${getFormattedTime(new Date(Date.now()))}`,
+    );
     const numFiles = req.body.numFiles;
     const value = req.body.value;
     function writeFile(fileNo, maxFileNo) {
@@ -43,7 +57,11 @@ app.post("/writeSeq", (req, res) => {
                 writeFile(fileNo + 1, maxFileNo);
             } else {
                 res.status(200).send("done");
-                console.log(`answered request nr. ${req.id} at ${getFormattedTime(new Date(Date.now()))}\n`);
+                console.log(
+                    `answered request nr. ${req.id} at ${getFormattedTime(
+                        new Date(Date.now()),
+                    )}\n`,
+                );
             }
         });
     }
@@ -51,26 +69,44 @@ app.post("/writeSeq", (req, res) => {
 });
 
 app.post("/writePar", (req, res) => {
-    console.log(`[WRITE-PAR] start processing received request nr. ${req.id} at ${getFormattedTime(new Date(Date.now()))}`);
+    console.log(
+        `[WRITE-PAR] start processing received request nr. ${
+            req.id
+        } at ${getFormattedTime(new Date(Date.now()))}`,
+    );
     let executions = [];
     const { numFiles, value } = req.body;
     for (let i = 0; i < numFiles; i++) {
-        executions.push(fs.promises.writeFile(`${dataDir}/file${i}.txt`, value));
+        executions.push(
+            fs.promises.writeFile(`${dataDir}/file${i}.txt`, value),
+        );
     }
     Promise.all(executions).then(() => {
         res.status(200).send("done");
-        console.log(`answered request nr. ${req.id} at ${getFormattedTime(new Date(Date.now()))}\n`);
+        console.log(
+            `answered request nr. ${req.id} at ${getFormattedTime(
+                new Date(Date.now()),
+            )}\n`,
+        );
     });
 });
 
 function getFormattedTime(dateTS) {
     let dateObj = new Date(dateTS);
-    return `${dateObj.getHours()}:${('00' + dateObj.getMinutes()).slice(-2)}:${('00' + dateObj.getSeconds()).slice(-2)}.${('000' + dateObj.getMilliseconds()).slice(-3)}`;
+    return `${dateObj.getHours()}:${("00" + dateObj.getMinutes()).slice(-2)}:${(
+        "00" + dateObj.getSeconds()
+    ).slice(-2)}.${("000" + dateObj.getMilliseconds()).slice(-3)}`;
 }
 
-async function getAverageExecutionTimeForEndpoint(endpointURL, executionNumber, numFiles = 5000, value = "Kendlbacher") {
+async function getAverageExecutionTimeForEndpoint(
+    endpointURL,
+    executionNumber,
+    numFiles = 5000,
+    value = "Kendlbacher",
+) {
     let executions = [];
-    if (fs.existsSync(dataDir)) throw new Error("Data directory already exists!");
+    if (fs.existsSync(dataDir))
+        throw new Error("Data directory already exists!");
 
     for (let i = 0; i < executionNumber; i++) {
         try {
@@ -87,7 +123,7 @@ async function getAverageExecutionTimeForEndpoint(endpointURL, executionNumber, 
             }),
             headers: {
                 "Content-Type": "application/json",
-            }
+            },
         });
         console.log(`Took ${Date.now() - startTime}ms\n`);
         executions.push(Date.now() - startTime);
@@ -100,16 +136,27 @@ try {
     app.listen(3000, async () => {
         console.log("Server running on http://localhost:3000/");
 
-        let normal = await getAverageExecutionTimeForEndpoint("http://localhost:3000/write", 3);
-        let sequential = await getAverageExecutionTimeForEndpoint("http://localhost:3000/writeSeq", 3);
-        let parallel = await getAverageExecutionTimeForEndpoint("http://localhost:3000/writePar", 3);
-
+        let normal = await getAverageExecutionTimeForEndpoint(
+            "http://localhost:3000/write",
+            3,
+        );
+        let sequential = await getAverageExecutionTimeForEndpoint(
+            "http://localhost:3000/writeSeq",
+            3,
+        );
+        let parallel = await getAverageExecutionTimeForEndpoint(
+            "http://localhost:3000/writePar",
+            3,
+        );
 
         console.log(`[NORMAL] Average execution time: ${normal.toFixed(2)}ms`);
-        console.log(`[SEQUENTIAL] Average execution time: ${sequential.toFixed(2)}ms`);
-        console.log(`[PARALLEL] Average execution time: ${parallel.toFixed(2)}ms`);
+        console.log(
+            `[SEQUENTIAL] Average execution time: ${sequential.toFixed(2)}ms`,
+        );
+        console.log(
+            `[PARALLEL] Average execution time: ${parallel.toFixed(2)}ms`,
+        );
     });
 } finally {
-    if (fs.existsSync(dataDir))
-        fs.rmSync(dataDir, { recursive: true });
+    if (fs.existsSync(dataDir)) fs.rmSync(dataDir, { recursive: true });
 }
